lab_2/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.0
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.143
litellm==1.51.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.1.0
mixpanel==4.10.1
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.53.0
orjson==3.10.11
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
posthog==3.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pypandoc==1.14
pyperclip==1.9.0
python-dateutil
python-dotenv==1.0.1
python-telegram-bot==21.6
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.1
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.19.1
tqdm==4.66.6
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.17.1
zipp==3.20.2
langchain-community
PyPDF2
sentence-transformers